requirements.txt
bcrypt==4.0.1
oracledb>=2.5.0,<3.0.0
python-dotenv>=1.0.0,<1.1.0
requests==2.32.4

API_BASE_URL=https://mindicador.cl/api
